Migrate 'build' and 'compile' scripts to Node
