--- catboost/node-package/build.sh
@@ -1,3 +1,0 @@
-#!/usr/bin/env sh
-
-node ./build_scripts/bootstrap.js build_model && ./compile.sh
