+++ catboost/node-package/build_scripts/build_model.ts
@@ -1,25 +1,3 @@
-import {execProcess} from './common';
-
-async function buildModelInterfaceLibrary() {
-    const srcPath = process.env['CATBOOST_SRC_PATH'] || '../..';
-    const result = await execProcess(`${srcPath}/ya make -r ${srcPath}/catboost/libs/model_interface -o ./build`);
-    if (result.code !== 0) {
-        console.error(`Building catboostmodel library failed: 
-            ${result.code} ${result.signal} ${result.err?.message}`);
-        process.exit(1);
-    }
-}
-
-async function buildModel() {
-    const result = await execProcess('node-gyp configure');
-    if (result.code !== 0) {
-        console.error(`Node-gyp configuration failed: 
-${result.code} ${result.signal} ${result.err?.message}`);
-        process.exit(1);
-    }
-
-    await buildModelInterfaceLibrary();
-    // TODO
-}
+import {buildModel} from './build';
 
 buildModel();
