+++ catboost/node-package/build_scripts/build_ya.ts
@@ -1,0 +1,8 @@
+import {buildModel, compileBindings} from './build';
+
+async function buildYa() {
+    await buildModel();
+    await compileBindings();
+}
+
+buildYa();
