+++ catboost/node-package/build_scripts/compile.ts
@@ -1,0 +1,4 @@
+import {compileBindings} from './build';
+
+
+compileBindings();
