--- catboost/node-package/compile.sh
@@ -1,5 +1,0 @@
-#!/usr/bin/env sh
-
-tsc && \
-    cp bindings/catboost.* lib && \
-    node-gyp build
